PATCH: [perl #133175] script run free from wrong pool panic
Setting the pointer to NULL after freeing signals the code in later
interations that it has been freed already

No test is added because it could become outdated (not testing what it
was designed to test) with a new Unicode version changing the underlying
data.  This bug was discovered by testing on Unicode 7.0, and the data
changed so that there was not a problem by Unicode 10.0.

(cherry picked from commit 77dddf94110be150e8cd1a4e9d469e418a788b8d)