mktables: Handle cjkiicore properly
This property is not normally compiled by perl, but an installation may
choose to use it.  It was failing some tests because this is a special
property that is like a perl dual-var.  It is both binary, and
non-binary, and commit 346f9bfbe12 forgot that.

(cherry picked from commit 954b558543c3a6941cc79943426589dcdfaadb7e)