regen/mk_invlists.pl: Fix chicken and egg problem
The problem here is that it was using a regular expression pattern to
determine if a code point is the integer 0.  When a new Unicode release
comes along and adds a new block of decimals, this routine should be run
before the interpreter is compiled for real.  And the pattern won't know
about the new block, so this would fail.

Solve the problem by using only Unicode::UCD to discover this info, and
not a pattern.

(cherry picked from commit a1c8344dcb15e1c52b83ca71795ec7c711bda53e)