partial revert "mingw: rename importlib to libcperl.a"
This reverts commit c27314b463b3b285217b651d581da5b43a9e26e3. The commit
message says only the filename of the mingw lib file was changed and MSV
must beleft as before. But in GNUmakefile and makefile.mk both the Mingw
and MSVC branches were renamed. This caused breakage in ../cpan/B-C/t/*.t
tests. The breakage was not caught by CI because all CI uses the nmake
Makefile to build themselves.