fix generating parallel win32 libcperl.a
reverts parts of "mingw: no broken dlltool"
commit 27d02a0323b5e0a527385f4dc103b676a3d44148. Removing dlltool for
making .dll files is okay since dlltool is obsolete in creating a dll but
that commit also removed dlltool for making a .a file from a .def file. A
parallel win32 build creates the .a file that is a binary description of
libperl.dll's interface from a .def file so XS modules can be built
before/parallel with libperl .o files being compiled and before/parallel
with linker creating libperl.dll. Not create libperl.a as a byproduct of
the linker creating libperl.dll. There were race random problems I
encountered in a "gmake -j2" build. Also target
"gmake ..\lib\CORE\libcperl.a" did not create a libcperl.a on disk which is
wrong.