Clarify your name and email *will* appear in AUTHORS unless asked:
Neil Bowers and Andreas Koenig spotted that, based on GDPR, we should
clarify to people that their name and email will appear in AUTHORS when
submitting a patch and that it can be removed when asked.

Neil contributed the phrasing.

(cherry picked from commit bb28885784a1aeae5e0c72076218f714945b0f17)