Fix uninitialized error in my_atof3()
This function, newly introduced in 5.29, by
6928bedc792ff80f0cb915460a7eacae25fa9bdd, is buggy due to my misreading
the man page for strtoflt128().  There was no man page on my system, and
the one on-line is very terse, and could be interpreted as doing what I
wanted, which is to have the second parameter on input point to the end
position in the input string beyond which the function is not to look.
But in fact the function is expecting a NUL-terminated string.

This commit creates such a string by copying the original when it isn't
NUL-terminated, before calling strtoflt128().

(cherry picked from commit d94e901a397f05e1d0030c5cfd535ef97a6808b6)