Make utf8_to_uvchr() safer
This function is deprecated because the API doesn't allow it to
determine the end of the input string, so it can read off the far end.
But I just realized that since many strings are NUL-terminated, so we
can forbid it from reading past the next NUL, and hence make it safe in
many cases.

(cherry picked from commit aa3c16bd709ef9b9c8c785af48f368e08f70c74b)