Prepare for Unicode 11.0
Unicode 11 has some new data files needed for it, and some changes in
the boundary rules that need to be accounted for.  This does all that
can be done without causing tests to fail.  The LB algorithm has
changed, and tests would fail if we included the code changes needed for
that change in this commit.  Instead those few lines will come as part
of the Unicode 11.0 commit.

(cherry picked from commit c0734505dda7c70c48ff2cc5f95bb57bcbe04ba0)