mathoms.c: Make safer a deprecated function
This function is deprecated because it is lacking a length parameter, so
malformed UTF-8 may cause it to read beyond the buffer.  This commit
causes it to not read beyond a NUL character, which makes it safe for
the common case that the input is a C string.

(cherry picked from commit c6734c35eefe26067112bce2d763f16bc9d99333)