Revise kludge for [perl #133136]
Lexical::SealRequireHints has a test that is looking for internal Perl
behavior that has changed, and the maintainer has been unresponsive.
Since this test is high in the cpan river, in order to get 5.28
released, a kludge was added that used the old behavior.  But that old
behavior is about to be completely removed, so this does a different
kludge that loads utf8.pm, and that is what the .t is actually
expecting.  This leaves us free to remove the behavior without affecting
the module.

(cherry picked from commit 5ac3601b965b993adcea616254df8de3c66f6ef6)