Move Unicode \p{} definitions to regcomp.c
These are only used in compiling patterns.  They previously were placed
in utf8.c because they are large, and there is a copy of regcomp.c in
ext/re, so they would have use twice the space.

This commit changes things so that they only are used and defined in
regcomp.c, (not re_comp.c) so that duplication does not occur.  They are
accessed only from one function, and that is also moved from utf8.c to
regcomp.c, only compiled in regcomp.c, and referred to as an external by
re_comp.c

I had to change the names of the table.  Previously they started with
'PL_' in case any got exposed, but globvar.t mindlessly assumes that any
such variables in the file regcomp.c are globals, and wrongly complains.
It was easier to just change the prefix to 'UNI_' instead.

A few tables are used in regexec.c, and are duplicated in re_exec.c.
Things could be adjusted so that only one copy is used.  I tried this,
but the tables are far more intertwined in regexec.c functions than
the ones changed in this commit, as only a single function accesses
these.  Thus doing this would be a lot harder, and the payback isn't all
that much.  I started work to make them EXTCONSTs, and then discovered
the intertwining, but left in that work, unused.

(cherry picked from commit cef721997e14497f2fbc4be17ab736ad7ddfda29)