Add shadow warning Subroutine masks new package
This is the other way round, when the outer method already exists,
and a new sub package or class is added.
This needs to check on every package or class creation for an existing
sub with the same name.
So far only done on explicit package creation via package or class,
not implicitly via a fully qualified symbol name, sub fullname {}.

Improve the implementation of the previous commit for overlong stashnames.