Make sv_setsv_flags accept an inversion list src
Inversion lists are a type of SV, but aren't really convertible to other
types, nor has there been a need to.  But it makes things more
convenient for a future commit to allow one to just clobber whatever is
in an SV and replace it with an inversion list.

Even after this commit, there are restrictions to what the SV can have
that gets clobbered.  But this is a step towards more generality

(cherry picked from commit 3d74616334d940df3e648ff7f770055ccdc8ef4e)