regcomp.c: Add some potential code that's #ifdef'd out
This is in case we ever need it.  This checks for portability in the
code points specified in user-defined properties.  Previously there was
a check, but I couldn't get a warning to trigger unless there was also
overflow.  So that means the pattern compile failed due to the overflow,
and the portability warning was superfluous.  But, one can have
non-portable code points without overflow; just the old method didn't
properly detect them.  If we do ever need to detect and report on them,
the code is mostly written and in this commit.

(cherry picked from commit de5659380e7c1c711fbe4cef1ee09c7460bd5f55)