Storable: enable >2GB AvFILL check on store_hook
077ae636940628 added an if (count > I32_MAX)
Too many references returned by STORABLE_freeze check, which was not
enabled, because count was only int, not SSize_t.
xav_fill was always SSize_t, so use that, which does not wrap around
on 2GB.

This is a security issue on 64bit cperl and perl5, as both allow
SSize_t arrays.