Fix ffi: Parse error panic: free from wrong pool
leak the prep_cif value in CvFFILIB, as we are not sure if its the external
libffi malloc'ed cif, or the temp. DynaLoader::dl_load_file scalar libhandle IV
used during dl_load_file. We could protect this with a temp. flag though
in S_find_native() and S_find_symbol().

Fixes [cperl #373], mostly observed in win64 threaded, with -DPERL_TRACK_MEMPOOL
failing when SafeFree the external libffi cif, and the header->interpreter is missing.