fix shared ffi string returns
previously we always shared the string ptr, and returned the new offset into the PV.
But this fails when freeing the string, as we never allocated the offset.

If pointing into our original string, a substring,
use the efficient OOK offset trick with the shared string.
Perl_sv_chop fails on readonly strings, so temp. turn that off.

If pointing outside, copy it.