regcomp.c: Generate new regnode for /[[:posix:]]/l
This follows on to the previous commit and generates code to use the new
regnode.  This allows some simplifications.  The determination of the
regnode is moved later in the function that does this; the code that
backed this out if we guessed wrong is excised.

(cherry picked from commit d156f5cb3264950617984f631f4023cd2482df47)