remove -ffi from miniperl, add ppmini, xsutilsmini
This is to simplify missing LD_LIBRARY_PATH decls for the toolchain
bootstrap.
And also to mimic -Uuseffi where extern sub decls still need to
be valid, just calls need to die. Source code need to be parsable
without eval "", a single if ($Config{useffi}) should be enough to
switch to a PP variant.

Unfortunately we need ppmini also, because we don't want to slow
down enterffi run-time with an extra exported Perl_ffi_call() into xsutils.