regcomp.h: Swap struct vs typedef
This struct has two names.  I previously left the less descriptive one
as the primary because of back compat issues.  But I now realize that
regcomp.h is only used in the core, so it's ok to swap for the better
name to be primary.

(cherry picked from commit ea461dd624d982da3d7216fd18fd3caca818c2bb)