Fix Unicode methods with -> syntax: keep WAS_UTF8
See [cperl #377].
A direct method call with :: stays utf8,
but a method call with -> strips the utf8 flags.
In cperl this leads to an invalid identifier error under strict names.

Føø::Bær::nèw() keeps the METHOP/ENTERSUB utf8 flag in the GV SVOP.  But
with Føø::Bær->nèw() the nèw is stored interim as shared pvn, which strips
off the utf8 to be stored in the strtab - newSVpvn_share: bytes_from_utf8
&is_utf8 - which leads to the conversion to the entersub gv without the utf8
flag.  The WAS_UTF8 flag is not stored.

Now at least the WAS_UTF8 flag is stored in the global string table, but
the GV itself has no UTF8 flag yet.