ffi :symbol(): use the prototype slot for the C name
A CvEXTERN does not need a prototype, only a sig.
We use the prototype slot for the deviating c-symbol name,
when it defers from the perl symbol name.

TODO check on extern sub foo() :symbol(_foo); both flags HASSIG
and SIGOP are set, to avoid getting into the ck_entersub proto checker.