getcwd_sv/PathTools: fix ENAMETOOLONG
On some older systems the OS failed on getcwd() with
File name too long >4096 even if the buffer is large enough,
eg debian-7.11 i386. Fix the test to check for $!

Fix the HAS_GET_CURRENT_DIR_NAME check, and the is_ENAMETOOLONG macro.

Needs to be backported to v5.28.1c