regcomp.c: Omit warning if error about to be raised
This commit changes the code to skip a warning when it knows an error is
about to happen.  Currently this doesn't matter, as the warning would be
emitted only in a later pass, and the error would actually happen first,
so the warning doesn't get output at all.  But future commits will
change that, so this commit is in preparation for that.

(cherry picked from commit 0a0ee7b8b14e71974adcfd6457c0d0a5486c28aa)