regcomp.c: Defer setting the OP variant of a regnode
This is in preparation for a future commit.  We have allocated space and
know that the regnode will be some time of ANYOF one.  But we don't need
to know which one precisely until later than we had been setting it.

(cherry picked from commit ebc61c6df0a1d53ff385289afdd2552c413ed11d)