Consolidate code into a single macro
If we die during the code generation phase, we set the regex SV to be
freed during cleanup.

This consolidates many of those instances into one macro, so that it can
be easily changed.  And instead of tieing it to the particular phase, we
clean up whenever that SV actually exists.  This requires initializing
it to NULL.

(cherry picked from commit ed68fe11532c7073e43da9e3cb5f6d4409c44c16)