regcomp.c: Move fcn call out of loop
The loop in this case is by a goto label, and the function determines if
there are runtime code blocks in the pattern.  That doesn't change if we
have to reparse, so the return from the function doesn't change, so we
only have to call it once.

(cherry picked from commit 690dd6eba70d6a431cbb661b449c25d4d4e0a029)