fix 'for reverse @array' bug on AIX
RT #133558

Due to what appears to be a compiler bug on AIX (or perhaps it's
undefined behaviour which happens to work on other platforms), this line
of code in pp_iter():

    inc = 1 - (PL_op->op_private & OPpITER_REVERSED);

was setting inc to 4294967295 rather than to the expected -1 (inc was a
64-bit signed long).

Fix it with a couple of judicious (IV) casts (which ought to be a NOOP).

(cherry picked from commit d6139ec4a9065ae249ab512398326a70dfb2fea2)

cperl: adjusted for the cperl for loop improvements.