regcomp.c: Test for having /u earlier for \p{}
When \p{} or \p{} are encountered during parsing, that indicates that
the pattern should be compiled not under /d, but under /u, as these are
Unicode constructs.  This commit moves the test for that to somewhat
earlier.  This saves only a little work currently, but in a future
commit it saves a lot more wasted work.

(cherry picked from commit a3ff97cf29e5b9d0f9c6bb019c9680b6f7575bd6)